DRY: extract log viewer toolbar into a component
